fix(reporting): kappa-derived assurance bounds + rho-doubledagger active-set disjunct (review follow-up, #1016)
Address yu2C's review on #1027:

- CheckValidatorIndex bounds the assurer index by len(kappa) (GP v0.8.0
  eq:xtassurances) -- the same kappa the signature check indexes --
  instead of the ValidatorsCount constant.
- The availability threshold is computed from the active sequence:
  3*count > 2*len(kappa) (GP eq:availableworkreports 'more than 2/3'),
  replacing the ValidatorsSuperMajority constant. Value-identical while
  |kappa| == V (tiny 5, full 683).
- FilterAvailableReports implements the |kappa| != |kappa'| clearing
  disjunct of eq:availassignmentspostassurancesdef. Dead code under the
  offenders-zeroed invariant, but spec-shaped.
- Genuinely variable |kappa| support (codec length pins, config
  plumbing, variable-kappa tests) is tracked in #1037.

Also fold TestEncodeAvailabilityAssignment_V080Guarantee into
encode_test.go next to the RefineContext layout test and drop the
standalone rho_v080_test.go (review suggestion).
